A file is created for each structural model, for better organization, and the approx crate is used in the tests
